+++ language/src/main/java/i4gl/exceptions/InvalidCastException.java
@@ -8,7 +8,7 @@ public class InvalidCastException extends I4GLRuntimeException {
      */
     private static final long serialVersionUID = 810218821002797877L;
 
-    public InvalidCastException(Object object, BaseType descriptor) {
-        super("Can not cast " + object.toString() + " to " + descriptor.toString());
+    public InvalidCastException(Object object, BaseType type) {
+        super("Can not cast " + object.toString() + " to " + type.toString());
     }
 }