+++ CHANGELOG.md
@@ -18,6 +18,9 @@ Version 0.3.0
 - Reimplement TextType, CharType and Char1Type as ArrayType descendants
 - Fix CaseNode now compares with nodes
 - Reimplement ReadRecordFieldNode and WriteRecordFieldNode with InteropLibrary
+- Fix sqlca not available at parse time
+- Add explicit casting
+- Add SMALLINT to arithmetic nodes
 
 
 Version 0.2.0